feat: Add bounding box coordinates for sweden and geojson reader
Add function that reads a geojson and returns a shapely object. Also add a quickly sketched geojson file around the swedish borders. Lastly add a function which takes the highest lat/lon and makes it into the northeastern corner and vice versa for the south western corner. These coordinates will be used for the API call.