feat: Add a function to partition the API search and download all data on Sweden
Since the search for the full bounding box of sweden was to large to download at once. A function was implemented that splits the full box into several partitions. Each of these will be downloaded and put into the directory nobil_json. Later these will be reorganised and put into one file.