feat: Add function to flatten metadata from nested to flat dict
Add a function to flatten a nested json dictionary to a simple dict.
Based on observations on the pattern of the API data. This works both for
'attr/st' and 'attr/conn' metadata. Also add some explainatory context
on how the data is structured and why this works. Lastly also add a function
to get_stations from a directory of similarly structured json files.