feat: Run the functions and save the data into a csv file
Runs the functions to get the stations, turn them into rows based on their metadata and turn them into a dataframe. Ensure that they get columns on lats and lons separately. Lastly filter the dataframe so we only have swedish chargingstations left. Then we save the dataframe as a CSV.