feat: Flask server - with DB and REST api
Add a flask server with a sqlite database that works using REST API.
In models.py the datatypes are described, i.e. the schemas for the
data that is going to be used. __init__.py is used in order to run
the whole folder as an application and therefore be able to import
functions and models across the application without circular imports.

App.py simply runs the app and has no other logic. Routes describes
and defines how a user can interact with the API. Basic CRUD operations
are available for 3 out of 5 datatypes, or tables. These are users,
private chargers and public chargers.

Lastly the database is populated with data that we as a group spoofed,
but that is essentially true to the type of data that will be there in
the future.